requirements.txt

discord.py
pymongo
python-dotenv
requests
discord.py
